+++ go/oeco-sdk/v2beta/CHANGELOG.md
@@ -2,11 +2,11 @@
 
 ### 🩹 Fixes
 
-- added additional error handling; preparing for batch handling ([0fa903f](https://github.com/openecosystems/ecosystem/commit/0fa903f))
+-   added additional error handling; preparing for batch handling ([0fa903f](https://github.com/openecosystems/ecosystem/commit/0fa903f))
 
 ### ❤️ Thank You
 
-- Dimy Jeannot
+-   Dimy Jeannot
 
 ## 0.20.4 (2025-09-21)